% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{df_dict}
\alias{df_dict}
\title{Insert data dictionary}
\usage{
df_dict(x = NULL)
}
\arguments{
\item{x}{a data frame}
}
\value{
Inserts a data dictionary template as a markdown list at the cursor
}
\description{
Insert data dictionary
}
\examples{
\dontrun{
df_dict(warpbreaks)
}
}
